% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIT_Estimator.R
\name{WIT}
\alias{WIT}
\title{Title WIT estimator with a predetermined initial points and a fixed tuning parameter lambda}
\usage{
WIT(D, Y, Z, beta_old, lambda, kappa = 0.5, error_c = 0.001, error_t = 1e-05)
}
\arguments{
\item{D}{is a nx1 vector means endogenous variables of interest}

\item{Y}{is a nx1 vector standards for responds}

\item{Z}{is a nxp vector means potential p-IVs that contains valid and invalid IVs}

\item{beta_old}{is the initial points we choose to use}

\item{lambda}{is the tuning lambda parameter}

\item{kappa}{is 1/rho in paper, take 0.5 for default.}

\item{error_c}{tolerance for contraction region, take 1e-3 for default.}

\item{error_t}{tolerance for tightening region, take 1e-5 for default.}
}
\value{
WIT estimate
}
\description{
Title WIT estimator with a predetermined initial points and a fixed tuning parameter lambda
}
