% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WIT_Estimator.R
\name{WIT_practice}
\alias{WIT_practice}
\title{Title WIT Estimator with automatically MCD tuning strategy.}
\usage{
WIT_practice(
  D,
  Y,
  Z,
  Lambda_Seq,
  ini_lam = 0.05,
  kappa = 2,
  error_c = 0.001,
  error_t = 1e-04,
  num_trail = 2
)
}
\arguments{
\item{D}{is a nx1 vector means endogenous variables of interest}

\item{Y}{is a nx1 vector standards for responds}

\item{Z}{is a nxp vector means potential p-IVs that contains valid and invalid IVs}

\item{Lambda_Seq}{is a tuning parameter sequence of lambda. Generally take sqrt(p/n)*(0.1,0.2,...,2)}

\item{ini_lam}{for grouping initial lambda, generally vary from 0.02 to 0.2.}

\item{kappa}{is 1/rho in paper, take 0.5 for default.}

\item{error_c}{tolerance for contraction region, take 1e-3 for default.}

\item{error_t}{tolerance for tightening region, take 1e-5 for default.}

\item{num_trail}{is upper bound of how many initials we may take}
}
\value{

}
\description{
Title WIT Estimator with automatically MCD tuning strategy.
}
